--- openstack/blockstorage/v2/snapshots/doc.go
@@ -1,5 +1,0 @@
-// Package snapshots provides information and interaction with snapshots in the
-// OpenStack Block Storage service. A snapshot is a point in time copy of the
-// data contained in an external storage volume, and can be controlled
-// programmatically.
-package snapshots
