--- openstack/blockstorage/v2/snapshots/testing/doc.go
@@ -1,2 +1,0 @@
-// snapshots_v2
-package testing
