--- openstack/blockstorage/v2/volumes/testing/doc.go
@@ -1,2 +1,0 @@
-// volumes_v2
-package testing
